improved isExternal marking and separation of lib and app for plugins

+++ core/db/db_helper.py
@@ -617,16 +617,17 @@ class DbHelper:
         names = [{'_id': i} for i in names if len(i) > 0]
         self.__package_names.insert(names)
 
-    def delete_static_entries(self, uuids):
+    def delete_static_entries(self, uuids, plugins_only=False):
         """
         Deletes database entries for anything related to static analysis
         """
-        res = self.__static_analysis_db.linkUrl.delete_many({"uuid": {"$in": uuids}})
-        logger.info("linkUrl - removed {}".format(res.deleted_count))
-        res = self.__static_analysis_db.permissionList.delete_many({"uuid": {"$in": uuids}})
-        logger.info("permissionList - removed {}".format(res.deleted_count))
-        res = self.__static_analysis_db.thirdPartyPackages.delete_many({"uuid": {"$in": uuids}})
-        logger.info("thirdPartyPackages - removed {}".format(res.deleted_count))
+        if not plugins_only:
+            res = self.__static_analysis_db.linkUrl.delete_many({"uuid": {"$in": uuids}})
+            logger.info("linkUrl - removed {}".format(res.deleted_count))
+            res = self.__static_analysis_db.permissionList.delete_many({"uuid": {"$in": uuids}})
+            logger.info("permissionList - removed {}".format(res.deleted_count))
+            res = self.__static_analysis_db.thirdPartyPackages.delete_many({"uuid": {"$in": uuids}})
+            logger.info("thirdPartyPackages - removed {}".format(res.deleted_count))
         res = self.__static_analysis_db.apkAnalyses.delete_many({"uuid": {"$in": uuids}})
         logger.info("apkAnalyses - removed {}".format(res.deleted_count))
         self.__apk_info.update_many({"uuid": {"$in": uuids}},
@@ -805,7 +806,7 @@ class DbHelper:
             self.third_party_queue = []
 
     def insert_permission_info (self, packageName, versioncode, filename, permission,
-                              is_external, dest, externalPackageName, src):
+                        is_external, triggered_by_code, externalPackageName):
         self.perm_info_queue.append(
             {
                 "packageName": packageName,
@@ -813,9 +814,8 @@ class DbHelper:
                 "filename": filename,
                 "permission": permission,
                 "isExternal": is_external,
-                "dest": dest,
-                "externalPackageName": externalPackageName,
-                "src": src
+                "triggeredByCode": triggered_by_code,
+                "externalPackageName": externalPackageName
             })
         #logger.info("{} perm_info".format(len(self.perm_info_queue)))
         if len(self.perm_info_queue) >= constants.FLUSH_LIM:
